requirements.txt
fastapi
uvicorn
chromadb
langchain
langchain-community
langchain-chroma
langchain-ollama
sentence-transformers
pydantic
python-multipart
numpy
tqdm
ollama
pypdf
unstructured
langchain-text-splitters
pathspec
gitpython
unidiff
streamlit
